refactor: deduplicate tools from agents-server into agents package
Move tool implementations (bash, read, write, edit, fetch_url, web_search,
schedules) from agents-server to agents package, removing duplicate code.
Tools are now exported from `@electric-ax/agents`. Remove unused dependencies
(jsdom, turndown, readability) from agents-server.
