+++ packages/agents-runtime/skills/designing-entities/references/review-checklist.md
@@ -60,11 +60,11 @@ Apply only when the entity `ctx.spawn("worker", ...)` — the Electric Agents se
 
 These apply to the app's `server.ts` / entry point, not to any individual entity. Flag them in phase 4 if the design relies on the feature.
 
-| #   | Rule                                                                                                                                                                                                                              | Why                                                                                                                                                                             |
-| --- | --------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
-| AW1 | If any entity spreads `...ctx.electricTools` and expects schedule tools, pass `createElectricTools: (ctx) => createScheduleTools(ctx)` to `createRuntimeHandler`. Import `createScheduleTools` from `@electric-ax/agents-server`. | Without it, `ctx.electricTools` is `[]`. The Electric Agents dev server wires this automatically; standalone apps must opt in.                                                  |
-| AW2 | The app process has `ANTHROPIC_API_KEY` (and any other provider keys) in its environment at boot — via `.env` file, shell export, or process manager.                                                                             | `agent.run()` calls the LLM provider. Missing API key throws inside the agent loop. If the handler doesn't catch, it crashes the wake (or the process). Check at boot and warn. |
-| S8  | `creationSchema` fields use `.default()` or `.optional()` where possible.                                                                                                                                                         | The Electric Agents UI spawns entities with no args — the server rejects with 422 if required fields are missing. Use defaults so entities can be spawned from the UI.          |
+| #   | Rule                                                                                                                                                                                                                       | Why                                                                                                                                                                             |
+| --- | -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
+| AW1 | If any entity spreads `...ctx.electricTools` and expects schedule tools, pass `createElectricTools: (ctx) => createScheduleTools(ctx)` to `createRuntimeHandler`. Import `createScheduleTools` from `@electric-ax/agents`. | Without it, `ctx.electricTools` is `[]`. The Electric Agents dev server wires this automatically; standalone apps must opt in.                                                  |
+| AW2 | The app process has `ANTHROPIC_API_KEY` (and any other provider keys) in its environment at boot — via `.env` file, shell export, or process manager.                                                                      | `agent.run()` calls the LLM provider. Missing API key throws inside the agent loop. If the handler doesn't catch, it crashes the wake (or the process). Check at boot and warn. |
+| S8  | `creationSchema` fields use `.default()` or `.optional()` where possible.                                                                                                                                                  | The Electric Agents UI spawns entities with no args — the server rejects with 422 if required fields are missing. Use defaults so entities can be spawned from the UI.          |
 
 ## Gotchas catalogue (rationale source)
 
@@ -84,7 +84,7 @@ These are the documented foot-guns — all universal checks above trace back to 
 12. **Handler that never ends.** Infinite loops get killed.
 13. **Required creationSchema fields + no-args UI spawn.** The Electric Agents UI spawns entities with no args. The server now rejects with 422 if required fields are missing. Use `.default()` or `.optional()` so entities can be spawned from the UI.
 14. **Secrets in worker prompts.** Interpolating `process.env.*` into a worker `systemPrompt` or `initialMessage` leaks it into the entity's persisted streams. Do authenticated fetches in the manager and pass the raw response to the worker as data; let built-in tools read their own keys from env.
-15. **`ctx.electricTools` empty in standalone apps.** Pass `createElectricTools: (ctx) => createScheduleTools(ctx)` to `createRuntimeHandler`. Import `createScheduleTools` from `@electric-ax/agents-server`.
+15. **`ctx.electricTools` empty in standalone apps.** Pass `createElectricTools: (ctx) => createScheduleTools(ctx)` to `createRuntimeHandler`. Import `createScheduleTools` from `@electric-ax/agents`.
 
 ## Output format (phase 4)
