+++ packages/agents-server-conformance-tests/src/electric-agents-tests.ts
@@ -1943,7 +1943,7 @@ export function runElectricAgentsConformanceTests(
     }
 
     test(`bash tool captures stdout and stderr`, async () => {
-      const { createBashTool } = await import(`../../agents-server/src/index`)
+      const { createBashTool } = await import(`../../agents/src/index`)
       const tool = createBashTool(`/tmp`)
       const result = await tool.execute(`test-tc`, {
         command: `echo "hello" && echo "error" >&2`,
@@ -1954,25 +1954,21 @@ export function runElectricAgentsConformanceTests(
     })
 
     test(`bash tool enforces timeout`, async () => {
-      const { createBashTool } = await import(`../../agents-server/src/index`)
+      const { createBashTool } = await import(`../../agents/src/index`)
       const tool = createBashTool(`/tmp`)
       const result = await tool.execute(`test-tc`, { command: `sleep 60` })
       expect(result.details.timedOut).toBe(true)
     }, 35_000)
 
     test(`read_file rejects paths outside working directory`, async () => {
-      const { createReadFileTool } = await import(
-        `../../agents-server/src/index`
-      )
+      const { createReadFileTool } = await import(`../../agents/src/index`)
       const tool = createReadFileTool(`/tmp/test-workdir`)
       const result = await tool.execute(`test-tc`, { path: `../../etc/passwd` })
       expect(firstText(result)).toContain(`outside the working directory`)
     })
 
     test(`read_file rejects binary files`, async () => {
-      const { createReadFileTool } = await import(
-        `../../agents-server/src/index`
-      )
+      const { createReadFileTool } = await import(`../../agents/src/index`)
       const fs = await import(`node:fs/promises`)
       const path = await import(`node:path`)
 
@@ -1990,9 +1986,7 @@ export function runElectricAgentsConformanceTests(
     })
 
     test(`read_file rejects oversized files`, async () => {
-      const { createReadFileTool } = await import(
-        `../../agents-server/src/index`
-      )
+      const { createReadFileTool } = await import(`../../agents/src/index`)
       const fs = await import(`node:fs/promises`)
       const path = await import(`node:path`)
 
@@ -2010,13 +2004,13 @@ export function runElectricAgentsConformanceTests(
     })
 
     test(`web_search tool has correct interface`, async () => {
-      const { braveSearchTool } = await import(`../../agents-server/src/index`)
+      const { braveSearchTool } = await import(`../../agents/src/index`)
       expect(braveSearchTool.name).toBe(`web_search`)
       expect(typeof braveSearchTool.execute).toBe(`function`)
     })
 
     test(`fetch_url tool has correct interface`, async () => {
-      const { fetchUrlTool } = await import(`../../agents-server/src/index`)
+      const { fetchUrlTool } = await import(`../../agents/src/index`)
       expect(fetchUrlTool.name).toBe(`fetch_url`)
       expect(typeof fetchUrlTool.execute).toBe(`function`)
     })
