+++ packages/agents-server/package.json
@@ -48,29 +48,23 @@
     "@electric-ax/agents-runtime": "workspace:*",
     "@electric-sql/client": "^1.5.14",
     "@mariozechner/pi-agent-core": "^0.70.2",
-    "@mozilla/readability": "^0.6.0",
     "@opentelemetry/api": "^1.9.1",
     "@sinclair/typebox": "^0.34.48",
     "ajv": "^8.18.0",
     "cron-parser": "^5.5.0",
     "drizzle-orm": "^0.44.0",
     "fastq": "^1.20.1",
-    "jsdom": "^28.1.0",
     "lmdb": "^3.5.1",
     "pino": "^10.3.1",
     "pino-pretty": "^13.0.0",
     "postgres": "^3.4.0",
-    "turndown": "^7.2.2",
-    "turndown-plugin-gfm": "^1.0.2",
     "undici": "^7.24.7"
   },
   "devDependencies": {
     "@electric-ax/agents": "workspace:*",
     "@electric-ax/agents-server-ui": "workspace:*",
     "@electric-ax/agents-server-conformance-tests": "workspace:*",
-    "@types/jsdom": "^27.0.0",
     "@types/node": "^22.19.15",
-    "@types/turndown": "^5.0.6",
     "@vitest/coverage-v8": "^4.1.0",
     "drizzle-kit": "^0.31.0",
     "tsdown": "^0.9.0",
