--- packages/agents-server/src/electric-agents/tools/turndown-plugin-gfm.d.ts
@@ -1,5 +1,0 @@
-declare module 'turndown-plugin-gfm' {
-  import type TurndownService from 'turndown'
-
-  export function gfm(service: TurndownService): void
-}
