+++ packages/agents-server/test/edit-tool-read-guard.test.ts
@@ -2,8 +2,8 @@ import { mkdtemp, rm, writeFile } from 'node:fs/promises'
 import { tmpdir } from 'node:os'
 import { join } from 'node:path'
 import { afterEach, beforeEach, describe, expect, it } from 'vitest'
-import { createEditTool } from '../src/electric-agents/tools/edit'
-import { createReadFileTool } from '../src/electric-agents/tools/read-file'
+import { createEditTool } from '../../agents/src/tools/edit'
+import { createReadFileTool } from '../../agents/src/tools/read-file'
 
 describe(`edit tool read-first guard`, () => {
   let cwd: string
