+++ packages/agents-server/test/write-tool.test.ts
@@ -2,7 +2,7 @@ import { mkdtemp, readFile, rm } from 'node:fs/promises'
 import { tmpdir } from 'node:os'
 import { join } from 'node:path'
 import { afterEach, beforeEach, describe, expect, it } from 'vitest'
-import { createWriteTool } from '../src/electric-agents/tools/write'
+import { createWriteTool } from '../../agents/src/tools/write'
 
 describe(`write tool`, () => {
   let cwd: string
